Replace `.setProperty` in favor of `.propertyName = value`
`CSSStyleDeclaration.setProperty()` calls accept the hyphen-case of the
CSS property name to be modified. Since we are forwarding the `style`
object of the JSX DOM element directly (which has camelCased names),
styles with more than one word in the property name won't be applied.

The `style.cssPropertyName = 'value';` variant does accept a camelCase
name, so replacing the former with the latter should fix the bug.

Closes #23.